add: starter framework
A c++ file provided by the instructor that reads a text file passed from
the command-line into an array. If the file extension is '.rle', the
data is passed to a filler decompress function that creates a '.plain'
file, else the data is passed to a filler compress function that creates
a '.rle' file.